Fix: 삭제 기능 멱등성 보장 및 로직 개선 (#53)
* [#52] Fix: 음식 중복 삭제 멱등성 보장을 위한 처리 및 테스트 추가

* [#52] Fix: 카테고리 삭제 시 음식 이동 및 멱등성 보장 로직 구현 및 테스트 추가

* [#52] Refactor: 카테고리 삭제 시 음식 이동 로직 및 FoodManager 의존성 제거

* [#52] Chore: CategoryRepository의 파라미터 타입 일관성있게 수정

+++ src/main/java/soon/fridgely/domain/category/service/CategoryRemover.java
@@ -5,20 +5,34 @@ import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 import soon.fridgely.domain.category.dto.command.DeleteCategory;
 import soon.fridgely.domain.category.entity.Category;
+import soon.fridgely.domain.category.repository.CategoryRepository;
 import soon.fridgely.domain.category.validator.CategoryValidator;
+import soon.fridgely.domain.food.service.FoodManager;
+import soon.fridgely.global.support.exception.CoreException;
+import soon.fridgely.global.support.exception.ErrorType;
 
 @RequiredArgsConstructor
 @Component
 public class CategoryRemover {
 
     private final CategoryValidator categoryValidator;
-    private final CategoryFinder categoryFinder;
+    private final FoodManager foodManager;
+    private final CategoryRepository categoryRepository;
 
     @Transactional
     public void remove(DeleteCategory deleteCategory) {
-        Category category = categoryFinder.findByRefrigerator(deleteCategory.categoryId(), deleteCategory.refrigeratorId());
+        Category category = categoryRepository.findByIdAndRefrigeratorId(deleteCategory.categoryId(), deleteCategory.refrigeratorId())
+            .orElseThrow(() -> new CoreException(ErrorType.NOT_FOUND_DATA));
+        if (category.isDeleted()) {
+            return;
+        }
+
         categoryValidator.validateNotDefaultType(category);
+
+        foodManager.moveAllFoodsToFallback(deleteCategory.refrigeratorId(), category.getId());
+
         category.delete();
+        categoryRepository.save(category); // foodManager에서 영속성 컨텍스트가 초기화되므로 다시 저장 필요
     }
 
 }
